requirements.txt
altair==5.5.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
easyocr==1.7.2
filelock==3.20.0
Flask==3.1.2
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.11.0
networkx==3.5
ninja==1.13.0
nltk==3.9.2
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pillow==12.0.0
protobuf==6.33.0
pyarrow==21.0.0
pyclipper==1.3.0.post6
pycparser==2.23
pydeck==0.9.1
pyphen==0.17.2
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytrends==4.9.2
pytz==2025.2
PyYAML==6.0.3
rake-nltk==1.0.6
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.28.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
textstat==0.7.11
threadpoolctl==3.6.0
tifffile==2025.10.16
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
